harden error paths and fix resource leaks across codebase
Systematic robustness review covering 30 source files:

- Add NULL checks after malloc/strdup/calloc (agents, cmd_tool, journal,
  llm, memory, prompt_optimize, provider, provider_anthropic, tui,
  ui_md_gen)
- Fix memory leaks: tool_file (resolved on error), tool_subtask
  (recalled_keys, fire_ledger), workspace (tags in dedup), commands
  (pbs outside if block)
- Fix mutex unlock before early return (memory_build_index/listing)
- Fix buffer overflows: md_html (parse_table_cells bounds), playbook
  (popen_cmd undersize), llm (repair_json +2)
- Fix logic bugs: regression (double-counting errors), matrix (wrong
  content variable for image path), store (retry short writes),
  tool_memory (reject bad consolidation responses), tool_todo
  (orig_count on error free), workspace (reindex before delete)
- Fix off-by-one in html_extract href scan loop
- Fix timeout overflow in mailbox (use time_t arithmetic)
- Add bounds-check on confidence index (commands, tool_session)
- Clamp UTF-8 boundaries (tool_session, tui)
- Use size_t instead of int for sizes (subprocess, toml)
- Skip cycling detection on OOM instead of UB (react)
- Use global memory for embeddings in workspace context (react_eviction)
- Add thread-safety comment on static buffer (ui_nav)

+++ src/subprocess.c
@@ -128,7 +128,7 @@ subprocess_result_t subprocess_run(char *const argv[],
         }
 
         /* Check caps */
-        if (max_bytes > 0 && (int)out->len >= max_bytes) {
+        if (max_bytes > 0 && out->len >= (size_t)max_bytes) {
             r.output_capped = 1;
             break;
         }
@@ -147,7 +147,7 @@ subprocess_result_t subprocess_run(char *const argv[],
             if (n <= 0) break;  /* EOF or error */
 
             /* Enforce byte cap with partial write */
-            if (max_bytes > 0 && (int)(out->len + (size_t)n) > max_bytes) {
+            if (max_bytes > 0 && (out->len + (size_t)n) > (size_t)max_bytes) {
                 size_t remaining = (size_t)max_bytes - out->len;
                 if (remaining > 0) str_append(out, buf, remaining);
                 /* Count lines in the accepted portion */
@@ -171,7 +171,7 @@ subprocess_result_t subprocess_run(char *const argv[],
             char buf[NASH_PATH_MAX];
             ssize_t n;
             while ((n = read(pipefd[0], buf, sizeof(buf))) > 0) {
-                if (max_bytes > 0 && (int)(out->len + (size_t)n) > max_bytes) {
+                if (max_bytes > 0 && (out->len + (size_t)n) > (size_t)max_bytes) {
                     size_t remaining = (size_t)max_bytes - out->len;
                     if (remaining > 0) str_append(out, buf, remaining);
                     break;
@@ -190,7 +190,7 @@ subprocess_result_t subprocess_run(char *const argv[],
                 char buf[NASH_PATH_MAX];
                 ssize_t n;
                 while ((n = read(pipefd[0], buf, sizeof(buf))) > 0) {
-                    if (max_bytes > 0 && (int)(out->len + (size_t)n) > max_bytes) {
+                    if (max_bytes > 0 && (out->len + (size_t)n) > (size_t)max_bytes) {
                         size_t remaining = (size_t)max_bytes - out->len;
                         if (remaining > 0) str_append(out, buf, remaining);
                         break;
